requirements.txt
Flask==2.3.3
python-dotenv
xhtml2pdf
openai
